+++ NEWS.md
@@ -4,7 +4,7 @@
 * Microscopic residual disease is regarded as no residual disease
 * Add residual disease condition for non-endometrioids in ESMO 2016 "high" case
 * Remove non-endometrioid condition in ESMO 2020 "intermediate" case for Stage IA, no myometrial invasion
-* In ESMO 2020 "high" case Stage I-IVA, endometrioid, >0% myometrial invasion, p53abn, (no residual disease if known), allow residual disease to also be missing if stage is low (I-II). If high stage, residual disease must be explicitly none.
+* In all ESMO 2020 risk groups involving residual disease, allow residual disease to also be missing if stage is low (I-II). If high stage, residual disease must be explicitly none.
 * Return value of all `assign_esmo*()` functions are now factors with levels appropriately ordered
 
 # EndoTools 0.3.1